refactor: 상행, 하행 종점역이 등록되지 않은 역인 경우 예외를 던지는 기능 위치 변경
Line 객체의 역할로 보기에는 어렵다고 판단하여 LineService에서 유효성 검증을 진행하도록 변경
