+++ README.md
@@ -29,12 +29,12 @@
 
 ### 지하철 노선 관련 기능
 - [x] 지하철 노선을 등록하는 기능
-  - [x] 중복된 노선 이름을 등록하면 예외를 발생시킨다 
-  - [x] 지하철 노선 이름이 2글자 미만이면 예외를 던진다
+- [x] 중복된 노선 이름을 등록하면 예외를 발생시킨다 
+- [x] 지하철 노선 이름이 2글자 미만이면 예외를 던진다
 - [x] 지하철 노선을 삭제하는 기능
 - [x] 노선 등록 시 상행 종점역과 하행 종점역을 입력받는 기능
-  - [x] 상행, 하행 종점역이 등록되지 않은 역인 경우 예외를 던진다
-  - [x] 노선의 역 갯수 초기값이 2가 아닌 경우 예외를 던진다 
+- [x] 상행, 하행 종점역이 등록되지 않은 역인 경우 예외를 던진다
+- [x] 노선의 역 갯수 초기값이 2가 아닌 경우 예외를 던진다 
 - [x] 지하철 노선의 목록을 조회하는 기능 
 
 ### 지하철 구간 추가 기능
